+++ conan/api/subapi/download.py
@@ -1,7 +1,7 @@
 from conan.api.model import Remote
 from conan.api.subapi import api_method
+from conan.cli.command import ConanOutput
 from conan.cli.conan_app import ConanApp
-from conan._internal.output import ConanOutput
 from conans.client.source import retrieve_exports_sources
 from conans.errors import ConanException
 from conans.model.package_ref import PkgReference
