+++ conan/api/subapi/export.py
@@ -1,8 +1,8 @@
 import os
 
 from conan.api.subapi import api_method
+from conan.cli.command import ConanOutput
 from conan.cli.conan_app import ConanApp
-from conan._internal.output import ConanOutput
 from conans.client.cmd.export import cmd_export
 from conans.client.conanfile.package import run_package_method
 from conans.client.graph.graph import BINARY_INVALID
