+++ conan/cli/cli.py
@@ -8,15 +8,14 @@ from collections import defaultdict
 from difflib import get_close_matches
 from inspect import getmembers
 
-from conans import __version__ as client_version
 from conan.api.conan_api import ConanAPIV2
-from conan.cli.command import ConanSubCommand
+from conan.cli.command import ConanSubCommand, cli_out_write, ConanOutput, Color
 from conan.cli.exit_codes import SUCCESS, ERROR_MIGRATION, ERROR_GENERAL, USER_CTRL_C, \
     ERROR_SIGTERM, USER_CTRL_BREAK, ERROR_INVALID_CONFIGURATION, ERROR_INVALID_SYSTEM_REQUIREMENTS
-from conan._internal.output import ConanOutput, cli_out_write, Color
+from conans import __version__ as client_version
 from conans.client.cache.cache import ClientCache
-from conans.errors import ConanInvalidSystemRequirements
 from conans.errors import ConanException, ConanInvalidConfiguration, ConanMigrationError
+from conans.errors import ConanInvalidSystemRequirements
 from conans.util.files import exception_message_safe
