+++ conan/cli/command.py
@@ -1,8 +1,12 @@
 import argparse
+import sys
 import textwrap
 
+from colorama import Style
+
+import conan._internal.output
 from conan.cli.commands import add_log_level_args, process_log_level_args
-from conan._internal.output import cli_out_write
+from conans.client.userio import color_enabled
 from conans.errors import ConanException
 
 COMMAND_GROUPS = {
@@ -219,3 +223,20 @@ def conan_subcommand(formatters=None):
         return cmd
 
     return decorator
+
+
+def cli_out_write(data, fg=None, bg=None, endline="\n", indentation=0):
+    fg_ = fg or ''
+    bg_ = bg or ''
+    if color_enabled(sys.stdout):
+        data = f"{' ' * indentation}{fg_}{bg_}{data}{Style.RESET_ALL}{endline}"
+    else:
+        data = f"{' ' * indentation}{data}{endline}"
+
+    sys.stdout.write(data)
+
+
+# FIXME: This is a bit weird. We want a public ConanOutput but also it is used from many places
+#        in the internal implementation
+ConanOutput = conan._internal.output.ConanOutput
+Color = conan._internal.output.Color
