+++ conan/cli/commands/build.py
@@ -1,12 +1,11 @@
 import os
 
-from conan.cli.command import conan_command, COMMAND_GROUPS
+from conan.cli.command import conan_command, COMMAND_GROUPS, ConanOutput
 from conan.cli.commands import make_abs_path
 from conan.cli.commands.install import graph_compute, _get_conanfile_path
 from conan.cli.common import _add_common_install_arguments, _help_build_policies, \
     get_multiple_remotes, add_lockfile_args, add_reference_args, save_lockfile_out
 from conan.cli.conan_app import ConanApp
-from conan._internal.output import ConanOutput
 from conans.client.conanfile.build import run_build_method
