+++ conan/cli/commands/config.py
@@ -1,6 +1,5 @@
-from conan.cli.command import conan_command, COMMAND_GROUPS, conan_subcommand
+from conan.cli.command import conan_command, COMMAND_GROUPS, conan_subcommand,  ConanOutput, Color
 from conan.cli.commands import json_formatter
-from conan._internal.output import ConanOutput, Color
 from conans.model.conf import BUILT_IN_CONFS
 from conans.util.config_parser import get_bool_from_text
