+++ conan/cli/commands/editable.py
@@ -1,10 +1,9 @@
 import os
 
-from conan.cli.command import conan_command, COMMAND_GROUPS, conan_subcommand
+from conan.cli.command import conan_command, COMMAND_GROUPS, conan_subcommand, ConanOutput
 from conan.cli.commands import make_abs_path
 from conan.cli.commands.install import _get_conanfile_path
 from conan.cli.conan_app import ConanApp
-from conan._internal.output import ConanOutput
 from conans.model.recipe_ref import RecipeReference
