+++ conan/cli/commands/export.py
@@ -1,10 +1,9 @@
 import json
 import os
 
-from conan.cli.command import conan_command, COMMAND_GROUPS, OnceArgument
+from conan.cli.command import conan_command, COMMAND_GROUPS, OnceArgument, ConanOutput
 from conan.cli.commands.install import _get_conanfile_path
 from conan.cli.common import get_lockfile, add_reference_args
-from conan._internal.output import ConanOutput
 
 
 def common_args_export(parser):
