+++ conan/cli/commands/inspect.py
@@ -1,8 +1,7 @@
 import json
 import inspect as python_inspect
 
-from conan.cli.command import conan_command, COMMAND_GROUPS, conan_subcommand
-from conan._internal.output import ConanOutput
+from conan.cli.command import conan_command, COMMAND_GROUPS, conan_subcommand, ConanOutput
 
 
 def _inspect_json_formatter(data):
