+++ conan/cli/commands/new.py
@@ -3,8 +3,7 @@ import shutil
 
 from jinja2 import Environment, meta, exceptions
 
-from conan.cli.command import conan_command, COMMAND_GROUPS, Extender
-from conan._internal.output import ConanOutput
+from conan.cli.command import conan_command, COMMAND_GROUPS, Extender, ConanOutput
 from conans.errors import ConanException
 from conans.util.files import save
