+++ conan/cli/common.py
@@ -1,8 +1,7 @@
 import os
 
-from conan.cli.command import Extender, OnceArgument, ExtenderValueRequired
+from conan.cli.command import Extender, OnceArgument, ExtenderValueRequired, ConanOutput
 from conan.cli.commands import make_abs_path
-from conan._internal.output import ConanOutput
 from conans.errors import ConanException
 from conans.model.graph_lock import LOCKFILE, Lockfile
